Changed lv_dropdown_add_option parameter from int to uint16_t

+++ src/lv_widgets/lv_dropdown.h
@@ -119,10 +119,10 @@ void lv_dropdown_set_static_options(lv_obj_t * ddlist, const char * options);
 /**
  * Add an options to a drop down list from a string.  Only works for dynamic options.
  * @param ddlist pointer to drop down list object
- * @param options a string without '\n'. E.g. "Four"
- * @param position the insert position, indexed from 0, LV_DROPDOWN_POS_LAST = end of string
+ * @param option a string without '\n'. E.g. "Four"
+ * @param pos  the insert position, indexed from 0, LV_DROPDOWN_POS_LAST = end of string
  */
-void lv_dropdown_add_option(lv_obj_t * ddlist, const char * option, int pos);
+void lv_dropdown_add_option(lv_obj_t * ddlist, const char * option, uint16_t pos);
     
 /**
  * Set the selected option
